[*]简化数据化 规范代码

+++ src/Practice/CozyLauncher.Plugin.Guide/Template/DataTemplate.cs
@@ -13,7 +13,7 @@ namespace CozyLauncher.Plugin.Guide.Template
     {
         public ITemplateInfo Info { get; set; }
 
-        public override UIElement InitTemplate(int width)
+        public override UIElement InitTemplate(double width)
         {
             return Info.GetInfoObject(width);
         }
